SinemonBackend/requirements.txt
django-sslserver==0.22
django-phonenumber-field==8.0.0
djangorestframework==3.15.1
dj-rest-auth==7.0.0
django-filter==23.5
django-cors-headers==4.5.0
django-rest-multiple-models==2.1.3
psycopg2-binary==2.9.10
django-allauth==0.63.6
django-countries==7.6.1
django-polymorphic==3.1.0
django-enumfield==3.1
django==3.2.12
numpy==2.0.2
pika==1.3.2
Pillow==11.2.1
uvicorn==0.34.2
python-socketio==5.13.0
asyncio==3.4.3
django-modeladmin-reorder==0.3.1
websockets==15.0.1
django-ckeditor==6.7.2
# uwsgi #keep commented if using windows.
python-dotenv==1.1.0
pycountry==24.6.1
pandas==2.2.3
